[NRH-15] add working stripe elements checkout widget to frontend
