+++ spa/src/components/MainRoutes.js
@@ -1,17 +1,23 @@
-import { BrowserRouter } from 'react-router-dom';
-import ProtectedRoute from 'components/authentication/ProtectedRoute';
+import { BrowserRouter } from "react-router-dom";
+import ProtectedRoute from "components/authentication/ProtectedRoute";
 
 // Slugs
-import { DASHBOARD_SLUG } from 'routes';
+import { DASHBOARD_SLUG } from "routes";
 
 // Children
-import Dashboard from 'components/dashboard/Dashboard';
+import Dashboard from "components/dashboard/Dashboard";
+
+// TEMP
+import TemporaryStripeCheckoutTest from "components/TEMP/TemporaryStripeCheckoutTest";
 
 function MainRoutes() {
   return (
     <BrowserRouter>
-      <ProtectedRoute path={DASHBOARD_SLUG}>
+      <ProtectedRoute exact path={DASHBOARD_SLUG}>
         <Dashboard />
+      </ProtectedRoute>
+      <ProtectedRoute path="/temp-pretend-payment">
+        <TemporaryStripeCheckoutTest />
       </ProtectedRoute>
     </BrowserRouter>
   );
