+++ spa/src/components/TEMP/TemporaryStripeCheckoutTest.styled.js
@@ -1,0 +1,22 @@
+import styled from "styled-components";
+
+export const TemporaryStripeCheckoutTest = styled.div`
+  display: flex;
+  flex-direction: column;
+  flex: 1;
+  justify-content: center;
+  align-items: center;
+`;
+
+export const Wrapper = styled.div`
+  width: 500px;
+`;
+
+export const FormStyled = styled.form`
+  background: purple;
+  padding: 2rem;
+  border-radius: 20px;
+  button {
+    margin-top: 1rem;
+  }
+`;
