+++ spa/src/components/authentication/util.js
@@ -1,6 +1,7 @@
 import { LS_USER, LS_CSRF_TOKEN } from "constants/authConstants";
 
 export function handleLoginSuccess(data) {
+  console.log("user org: ", data.user.organization);
   localStorage.setItem(LS_USER, JSON.stringify(data.user));
   localStorage.setItem(LS_CSRF_TOKEN, data.csrftoken);
 }
