--- spa/src/components/dashboard/TemporaryStripeCheckoutTest.styled.js
@@ -1,5 +1,0 @@
-import styled from "styled-components";
-
-export const TemporaryStripeCheckoutTest = styled.div`
-  width: 500px;
-`;
